// src/api/model/media.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * MaCarte API
 * Documentation OpenAPI de l'API MaCarte
 * OpenAPI spec version: 1.0.0
 */

/**
 * media
 */
export interface Media {
  /** identifiant du media */
  id?: number;
  /** nom de l'image originale */
  name?: string;
  /** nom du fichier enregistré */
  file?: string;
  /** username de l'utilisateur qui a chargé le media */
  owner?: string;
  /** id de l'organisation à qui appartient le media */
  organization_id?: string;
  /** nom de l'organisation à qui appartient le media */
  organization?: string;
  /** date du chargement de l'image */
  uploaded_at?: string;
  /** Taille de l'image (en octet) */
  size?: number;
  /** Nom du fichier contenant le media */
  folder?: string;
  /** Validité de l'image */
  valid?: boolean;
  /** Url de l'image */
  view_url?: string;
  /** Url de la vignette de l'image */
  thumb_url?: string;
}